requirements.txt
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.9.0
Django==4.2.9
django-cors-headers==4.3.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
MouseInfo==0.1.3
openai==1.10.0
outcome==1.3.0.post0
packaging==23.2
pillow==10.2.0
PyAutoGUI==0.9.54
pydantic==2.5.3
pydantic_core==2.14.6
pyee==8.2.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
#pyobjc-core==10.1
#pyobjc-framework-Cocoa==10.1
#pyobjc-framework-Quartz==10.1
pyperclip==1.8.2
pyppeteer==1.0.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dotenv
pytweening==1.0.7
pytz==2023.3.post1
requests==2.31.0
rubicon-objc==0.4.7
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3
websockets==10.4
wsproto==1.2.0
zipp==3.17.0
langchain[llms]